move dummy class only used for tests to the using test
